fix format issue and comment workflows
