+++ .github/workflows/update-model-info.yml
@@ -1,29 +1,29 @@
-name: update-model-info
+# name: update-model-info
 
-on:
-  pull_request:
-    # branches:
-    #   - dev
-    # types: [closed]
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
+# on:
+#   pull_request:
+#     # branches:
+#     #   - dev
+#     # types: [closed]
+#   # Allows you to run this workflow manually from the Actions tab
+#   workflow_dispatch:
 
-jobs:
-  build:
-    # if: github.event.pull_request.merged == true
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-      with:
-        fetch-depth: 0
-        lfs: 'true'
-    - name: update model info
-      run: |
-        python -m pip install -r requirements.txt
-        python $(pwd)/.github/update_model_info.py -p $(pwd)/models/
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        PR_NUMBER: ${{ github.event.number }}
-        token: ${{ secrets.UPDATE_MODEL_TOKEN }}
-        username: ${{ secrets.UPDATE_MODEL_USERNAME }}
-        email: ${{ secrets.UPDATE_MODEL_EMAIL }}
+# jobs:
+#   build:
+#     # if: github.event.pull_request.merged == true
+#     runs-on: ubuntu-latest
+#     steps:
+#     - uses: actions/checkout@v2
+#       with:
+#         fetch-depth: 0
+#         lfs: 'true'
+#     - name: update model info
+#       run: |
+#         python -m pip install -r requirements.txt
+#         python $(pwd)/.github/update_model_info.py -p $(pwd)/models/
+#       env:
+#         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+#         PR_NUMBER: ${{ github.event.number }}
+#         token: ${{ secrets.UPDATE_MODEL_TOKEN }}
+#         username: ${{ secrets.UPDATE_MODEL_USERNAME }}
+#         email: ${{ secrets.UPDATE_MODEL_EMAIL }}
